+++ src/nhc2-platform.ts
@@ -16,6 +16,8 @@ import { Device } from "nhc2-hobby-api/lib/event/device";
 import { Event } from "nhc2-hobby-api/lib/event/event";
 import { NHC2 } from "nhc2-hobby-api/lib/NHC2";
 
+import { NHC2Logger } from "./nhc2-logger";
+
 const PLUGIN_NAME = "homebridge-nhc2";
 const PLATFORM_NAME = "NHC2";
 
@@ -35,13 +37,24 @@ class NHC2Platform implements DynamicPlatformPlugin {
     .Characteristic;
 
   private readonly accessories: PlatformAccessory[] = [];
+  private readonly suppressedAccessories: string[] = [];
   private readonly nhc2: NHC2;
 
+  private readonly log: NHC2Logger;
+
   constructor(
-    private log: Logging,
+    private logger: Logging,
     private config: PlatformConfig,
     private api: API,
   ) {
+    this.log = new NHC2Logger(logger, config);
+    this.suppressedAccessories = config.suppressedAccessories || [];
+    if (this.suppressedAccessories) {
+      this.log.info("Suppressing accessories: ");
+      this.suppressedAccessories.forEach(acc => {
+        this.log.info("  - " + acc);
+      });
+    }
     this.nhc2 = new NHC2("mqtts://" + config.host, {
       port: config.port || 8884,
       clientId: config.clientId || "NHC2-homebridge",
@@ -50,16 +63,15 @@ class NHC2Platform implements DynamicPlatformPlugin {
       rejectUnauthorized: false,
     });
 
-    log.info("NHC2Platform finished initializing!");
+    this.log.info("NHC2Platform finished initializing!");
 
     api.on(APIEvent.DID_FINISH_LAUNCHING, async () => {
-      log.info("NHC2Platform 'didFinishLaunching'");
+      this.log.info("NHC2Platform 'didFinishLaunching'");
 
       await this.nhc2.subscribe();
       const nhc2Accessories = await this.nhc2.getAccessories();
-      this.addLights(nhc2Accessories);
-      this.addDimmers(nhc2Accessories);
-      this.addGenericSwitches(nhc2Accessories);
+      this.log.verbose("got " + nhc2Accessories.length + " accessories");
+      this.addAccessories(nhc2Accessories);
 
       this.nhc2.getEvents().subscribe(event => {
         this.processEvent(event);
@@ -72,7 +84,7 @@ class NHC2Platform implements DynamicPlatformPlugin {
   }
 
   public processEvent = (event: Event) => {
-    if(!!event.Params) {
+    if (!!event.Params) {
       event.Params.flatMap(param =>
         param.Devices.forEach((device: Device) => {
           const deviceAccessoryForEvent = this.findAccessoryDevice(device);
@@ -90,49 +102,45 @@ class NHC2Platform implements DynamicPlatformPlugin {
     return this.accessories.find(accessory => accessory.UUID === device.Uuid);
   }
 
-  private addLights(accessories: Device[]) {
-    const lights = accessories.filter(light => light.Model === "light");
-    lights.forEach(light => {
-      const newAccessory = new Accessory(light.Name as string, light.Uuid);
+  private addAccessories(accessories: Device[]) {
+    const mapping: { [index: string]: any } = {
+      light: {
+        service: this.Service.Lightbulb,
+        handlers: [this.addStatusChangeCharacteristic],
+      },
+      dimmer: {
+        service: this.Service.Lightbulb,
+        handlers: [
+          this.addStatusChangeCharacteristic,
+          this.addBrightnessChangeCharacteristic,
+        ],
+      },
+      socket: {
+        service: this.Service.Outlet,
+        handlers: [this.addStatusChangeCharacteristic],
+      },
+      generic: {
+        service: this.Service.Switch,
+        handlers: [this.addTriggerCharacteristic],
+      },
+    };
 
-      const newService = new this.Service.Lightbulb(light.Name);
-      this.addStatusChangeCharacteristic(newService, newAccessory);
-      newAccessory.addService(newService);
-
-      this.processDeviceProperties(light, newService);
-
-      this.registerAccessory(newAccessory);
-    });
-  }
-
-  private addDimmers(accessories: Device[]) {
-    const dimmers = accessories.filter(light => light.Model === "dimmer");
-    dimmers.forEach(dimmer => {
-      const newAccessory = new Accessory(dimmer.Name as string, dimmer.Uuid);
-
-      const newService = new this.Service.Lightbulb(dimmer.Name);
-      this.addStatusChangeCharacteristic(newService, newAccessory);
-      this.addBrightnessChangeCharacteristic(newService, newAccessory);
-      newAccessory.addService(newService);
-
-      this.processDeviceProperties(dimmer, newService);
-
-      this.registerAccessory(newAccessory);
-    });
-  }
-
-  private addGenericSwitches(accessories: Device[]) {
-    const genericSwitches = accessories.filter(acc => acc.Model === "generic");
-    genericSwitches.forEach(genericDevice => {
-      const newAccessory = new Accessory(genericDevice.Name as string, genericDevice.Uuid);
-
-      const newService = new this.Service.Switch(genericDevice.Name);
-      this.addTriggerCharacteristic(newService, newAccessory);
-      newAccessory.addService(newService);
-
-      this.processDeviceProperties(genericDevice, newService);
-
-      this.registerAccessory(newAccessory);
+    Object.keys(mapping).forEach(model => {
+      const config = mapping[model];
+      const accs = accessories.filter(
+        acc =>
+          !this.suppressedAccessories.includes(acc.Uuid) && acc.Model === model,
+      );
+      accs.forEach(acc => {
+        const newAccessory = new Accessory(acc.Name as string, acc.Uuid);
+        const newService = new config.service(acc.Name);
+        config.handlers.forEach((handler: any) => {
+          handler(newService, newAccessory);
+        });
+        newAccessory.addService(newService);
+        this.processDeviceProperties(acc, newService);
+        this.registerAccessory(newAccessory);
+      });
     });
   }
 
@@ -146,6 +154,13 @@ class NHC2Platform implements DynamicPlatformPlugin {
     this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [
       accessory,
     ]);
+    this.log.debug(
+      "registered accessory: " +
+        accessory.displayName +
+        " (" +
+        accessory.UUID +
+        ")",
+    );
   }
 
   private unregisterAccessory(accessory: PlatformAccessory) {
@@ -153,6 +168,13 @@ class NHC2Platform implements DynamicPlatformPlugin {
       accessory,
     ]);
     this.accessories.splice(this.accessories.indexOf(accessory), 1);
+    this.log.debug(
+      "unregistered accessory: " +
+        accessory.displayName +
+        " (" +
+        accessory.UUID +
+        ")",
+    );
   }
 
   private findExistingAccessory(newAccessory: PlatformAccessory) {
@@ -161,10 +183,10 @@ class NHC2Platform implements DynamicPlatformPlugin {
       .find(() => true);
   }
 
-  private addStatusChangeCharacteristic(
+  private addStatusChangeCharacteristic = (
     newService: Service,
     newAccessory: PlatformAccessory,
-  ) {
+  ) => {
     newService
       .getCharacteristic(this.Characteristic.On)
       .on(
@@ -177,12 +199,12 @@ class NHC2Platform implements DynamicPlatformPlugin {
           callback();
         },
       );
-  }
+  };
 
-  private addBrightnessChangeCharacteristic(
+  private addBrightnessChangeCharacteristic = (
     newService: Service,
     newAccessory: PlatformAccessory,
-  ) {
+  ) => {
     newService
       .getCharacteristic(this.Characteristic.Brightness)
       .on(
@@ -195,41 +217,30 @@ class NHC2Platform implements DynamicPlatformPlugin {
           callback();
         },
       );
-  }
+  };
 
-
-
-  private addTriggerCharacteristic(
+  private addTriggerCharacteristic = (
     newService: Service,
     newAccessory: PlatformAccessory,
-  ) {
+  ) => {
     newService
       .getCharacteristic(this.Characteristic.On)
       .on(
         CharacteristicEventTypes.SET,
         (value: CharacteristicValue, callback: CharacteristicSetCallback) => {
-          this.nhc2.sendTriggerBasicStateCommand(
-            newAccessory.UUID
-          );
+          this.nhc2.sendTriggerBasicStateCommand(newAccessory.UUID);
           callback();
         },
       );
-  }
+  };
 
   private processDeviceProperties(device: Device, service: Service) {
     if (!!device.Properties) {
-
       device.Properties.forEach(property => {
         if (property.Status === "On") {
           service.getCharacteristic(this.Characteristic.On).updateValue(true);
         }
         if (property.Status === "Off") {
-          service.getCharacteristic(this.Characteristic.On).updateValue(false);
-        }
-        if (property.BasicState === "On") {
-          service.getCharacteristic(this.Characteristic.On).updateValue(true);
-        }
-        if (property.BasicState === "Off") {
           service.getCharacteristic(this.Characteristic.On).updateValue(false);
         }
         if (!!property.Brightness) {
@@ -240,6 +251,4 @@ class NHC2Platform implements DynamicPlatformPlugin {
       });
     }
   }
-
-
 }